associate products in category

+++ ecommerce/src/main/java/com/example/ecommerce/service/CategoryService.java
@@ -1,16 +1,19 @@
 package com.example.ecommerce.service;
 
 import com.example.ecommerce.domain.Category;
+import com.example.ecommerce.domain.Product;
 import com.example.ecommerce.domain.dto.category.CategoryRequestDTO;
 import com.example.ecommerce.domain.dto.category.CategoryResponseDTO;
 import com.example.ecommerce.exception.BusinessException;
 import com.example.ecommerce.exception.ResourceNotFoundException;
 import com.example.ecommerce.mapper.CategoryMapper;
 import com.example.ecommerce.repository.CategoryRepository;
+import com.example.ecommerce.repository.ProductRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.PathVariable;
 
-import java.util.ArrayList;
+import java.util.HashSet;
 import java.util.List;
 
 @Service
@@ -19,10 +22,14 @@ public class CategoryService {
     private CategoryRepository categoryRepository;
 
     @Autowired
-    private CategoryMapper mapper;
+    private CategoryMapper categoryMapper;
+
+    @Autowired
+    private ProductRepository productRepository;
+
 
     public CategoryResponseDTO create(CategoryRequestDTO body) {
-        var category = mapper.toEntity(body);
+        var category = categoryMapper.toEntity(body);
 
         if (categoryRepository.existsByName(body.name())) {
             throw new BusinessException("Já existe uma categoria com esse nome");
@@ -30,7 +37,7 @@ public class CategoryService {
 
         categoryRepository.save(category);
 
-        return mapper.toResponseDTO(category);
+        return categoryMapper.toResponseDTO(category);
     }
 
     public CategoryResponseDTO getById(Long id) {
@@ -38,16 +45,25 @@ public class CategoryService {
                 () -> new ResourceNotFoundException("Categoria não encontrada!")
         );
 
-        return mapper.toResponseDTO(category);
+        return categoryMapper.toResponseDTO(category);
     }
 
     public List<CategoryResponseDTO> list() {
         List<Category> categories = categoryRepository.findAll();
 
-        return mapper.toDTOList(categories);
+        return categoryMapper.toDTOList(categories);
     }
 
     public void delete(Long id) {
+
+        var category = categoryRepository.findById(id).orElseThrow(
+                () -> new ResourceNotFoundException("Categoria não existe!")
+        );
+
+        if (category.getProducts() != null) {
+            throw new BusinessException("Não é possível deletar a categoria " +
+                    "pois ela possui produtos associados!");
+        }
         categoryRepository.deleteById(id);
     }
 
@@ -56,13 +72,78 @@ public class CategoryService {
                 () -> new ResourceNotFoundException("Categoria não encontrada!")
         );
 
-        mapper.updateEntityFromDTO(body, category);
+        categoryMapper.updateEntityFromDTO(body, category);
 
         var updatedCategory = categoryRepository.save(category);
 
-        return mapper.toResponseDTO(updatedCategory);
+        return categoryMapper.toResponseDTO(updatedCategory);
     }
 
-//    public associateProductInCategory
-//    public removeProductOfCategory
+    public void addProductToCategory(Product product, Category category) {
+        if (category.getProducts() == null) {
+            category.setProducts(new HashSet<>());
+        }
+
+        if (product.getCategories() == null) {
+            product.setCategories(new HashSet<>());
+        }
+
+        category.getProducts().add(product);
+        product.getCategories().add(category);
+    }
+
+    public CategoryResponseDTO associateProductInCategory(
+            @PathVariable Long idCategory,
+            @PathVariable Long idProduct
+    ) {
+        var product = productRepository.findById(idProduct).orElseThrow(
+                () -> new ResourceNotFoundException("Produto não encontrado!")
+        );
+
+        var category = categoryRepository.findById(idCategory).orElseThrow(
+                () -> new ResourceNotFoundException("Categoria não encontrada!")
+        );
+
+        this.addProductToCategory(product, category);
+
+        var updatedCategory = categoryRepository.save(category);
+
+        return categoryMapper.toResponseDTO(updatedCategory);
+    }
+
+    public void deleteProductToCategory(Product product, Category category) {
+        if (!category.getProducts().contains(product)) {
+            throw new ResourceNotFoundException("Não existe esse produto nessa categoria");
+        }
+
+        if (category.getProducts() != null) {
+            category.getProducts().remove(product);
+        }
+
+        if (product.getCategories() != null) {
+            product.getCategories().remove(category);
+        }
+
+        categoryRepository.save(category);
+        productRepository.save(product);
+    }
+
+    public CategoryResponseDTO removeProductOfCategory(
+            @PathVariable Long idCategory,
+            @PathVariable Long idProduct
+    ) {
+        var product = productRepository.findById(idProduct).orElseThrow(
+                () -> new ResourceNotFoundException("Produto não encontrado!")
+        );
+
+        var category = categoryRepository.findById(idCategory).orElseThrow(
+                () -> new ResourceNotFoundException("Categoria não encontrada!")
+        );
+
+        this.deleteProductToCategory(product, category);
+
+        var updatedCategory = categoryRepository.save(category);
+
+        return categoryMapper.toResponseDTO(updatedCategory);
+    }
 }
